Stockを整理
