--- resources/assets/front/components/organisms/StockItems/styles.module.scss
@@ -1,5 +1,0 @@
-
-.card {
-    display: flex;
-    flex-wrap: wrap
-}
